requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.10.1
pytz==2024.2
six==1.17.0
Werkzeug==3.1.3
requests
python-dotenv